api/requirements.txt
asgiref==3.6.0
Django==3.1.6
pytz==2022.7.1
sqlparse==0.4.3
djangorestframework
django-filter
